requirements.txt
# Core dependencies
httpx>=0.25.0
aiohttp>=3.8.0
pydantic>=2.4.0
pyyaml>=6.0.0
typer>=0.9.0
rich>=13.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Optional dependencies for future phases
# matplotlib>=3.7.0
# plotly>=5.15.0

